+++ engine/src/engine/search/negamax.rs
@@ -76,7 +76,8 @@ pub fn negamax(
 
     let mut previous_best_move: Option<Move> = None;
 
-    if let Some(tt_entry) = ctx.tt.get(game.zobrist, plies) {
+    let tt_entry = ctx.tt.get(game.zobrist, plies);
+    if let Some(ref tt_entry) = tt_entry {
         if !is_root && !is_pv && tt_entry.depth >= depth {
             let tt_score = tt_entry.score;
 
@@ -120,6 +121,7 @@ pub fn negamax(
                     game.zobrist,
                     tb_bound,
                     score,
+                    Eval::NONE,
                     depth,
                     ctx.tt.generation,
                     None,
@@ -135,7 +137,25 @@ pub fn negamax(
         }
     }
 
-    let eval = eval::eval(game);
+    let eval = match tt_entry {
+        Some(ref e) if e.eval != Eval::NONE => e.eval,
+        _ => {
+            let e = eval::eval(game);
+
+            ctx.tt.insert(
+                game.zobrist,
+                NodeBound::None,
+                Eval::NONE,
+                e,
+                0,
+                ctx.tt.generation,
+                None,
+                plies,
+            );
+
+            e
+        }
+    };
 
     if !is_root && !is_pv && !in_check {
         // Reverse futility pruning
@@ -282,6 +302,7 @@ pub fn negamax(
         game.zobrist,
         tt_node_bound,
         best_eval,
+        eval,
         depth,
         ctx.tt.generation,
         best_move,